rename: app 모듈 패키지 이동
