+++ app/src/main/java/com/techcourse/support/web/mapping/HandlerMappings.java
@@ -1,6 +1,5 @@
-package com.techcourse;
+package com.techcourse.support.web.mapping;
 
-import com.techcourse.mapping.ManualHandlerMappingWrapper;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
