+++ app/src/main/java/com/techcourse/support/web/resolver/ViewResolvers.java
@@ -1,7 +1,5 @@
-package com.techcourse;
+package com.techcourse.support.web.resolver;
 
-import com.techcourse.view.resolver.JsonViewResolver;
-import com.techcourse.view.resolver.JspViewResolver;
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.ArrayList;
 import java.util.List;
