+++ app/src/test/java/com/techcourse/support/web/mapping/ManualHandlerMappingTest.java
@@ -1,9 +1,8 @@
-package com.techcourse;
+package com.techcourse.support.web.mapping;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
 
-import com.techcourse.mapping.ManualHandlerMapping;
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
