+++ app/src/test/java/com/techcourse/support/web/resolver/JsonViewResolverTest.java
@@ -1,4 +1,4 @@
-package com.techcourse.view.resolver;
+package com.techcourse.support.web.resolver;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
